+++ packages/conan/recipes/ptex/conanfile.py
@@ -1,14 +1,18 @@
+# Copyright (c) Contributors to the conan-center-index Project. All rights reserved.
+# Copyright (c) Contributors to the aswf-docker Project. All rights reserved.
+# SPDX-License-Identifier: MIT
+
 from conans import ConanFile, tools, CMake
 import os
 
 required_conan_version = ">=1.38.0"
 
 
-class MaterialXConan(ConanFile):
-    name = "materialx"
-    description = "MaterialX is an open standard for the exchange of rich material and look-development content across applications and renderers."
-    topics = "conan", "materialx", "python", "vfx"
-    homepage = "https://github.com/AcademySoftwareFoundation/MaterialX"
+class PtexConan(ConanFile):
+    name = "ptex"
+    description = "Per-Face Texture Mapping for Production Rendering."
+    topics = "conan", "ptex", "vfx"
+    homepage = "https://github.com/wdas/ptex"
     license = "BSD-3-Clause"
     url = "https://github.com/AcademySoftwareFoundation/aswf-docker"
     settings = (
@@ -16,7 +20,6 @@ class MaterialXConan(ConanFile):
         "arch",
         "compiler",
         "build_type",
-        "python",
     )
     options = {
         "shared": [True, False],
@@ -32,11 +35,7 @@ class MaterialXConan(ConanFile):
     _source_subfolder = "source_subfolder"
 
     def requirements(self):
-        self.requires(
-            f"python/{os.environ['ASWF_PYTHON_VERSION']}@{self.user}/{self.channel}"
-        )
-        # Use vendored pybind11 for now
-        # self.requires(f"pybind11/{os.environ['ASWF_PYBIND11_VERSION']}@{self.user}/{self.channel}")
+        pass
 
     def build_requirements(self):
         self.build_requires(
@@ -44,10 +43,8 @@ class MaterialXConan(ConanFile):
         )
 
     def source(self):
-        tools.get(
-            f"https://github.com/AcademySoftwareFoundation/MaterialX/archive/v{self.version}.tar.gz"
-        )
-        os.rename(f"MaterialX-{self.version}", self._source_subfolder)
+        tools.get(f"https://github.com/wdas/ptex/archive/v{self.version}.tar.gz")
+        os.rename(f"ptex-{self.version}", self._source_subfolder)
 
     def _configure_cmake(self):
         if self._cmake:
@@ -58,12 +55,8 @@ class MaterialXConan(ConanFile):
 
         with tools.environment_append(tools.RunEnvironment(self).vars):
             self._cmake = CMake(self)
-            self._cmake.definitions["MATERIALX_BUILD_PYTHON"] = "ON"
-            self._cmake.definitions["MATERIALX_PYTHON_VERSION"] = os.environ[
-                "ASWF_PYTHON_VERSION"
-            ]
             if self.options.shared:
-                self._cmake.definitions["MATERIALX_BUILD_SHARED_LIBS"] = "ON"
+                self._cmake.definitions["BUILD_SHARED_LIBS"] = "1"
             self._cmake.configure(source_folder=self._source_subfolder)
             return self._cmake
 
@@ -77,10 +70,6 @@ class MaterialXConan(ConanFile):
         cmake.install()
 
     def package_info(self):
-        self.cpp_info.requires.append("python::PythonLibs")
-        # Use vendored pybind11 for now
-        # self.cpp_info.requires.append("pybind11::main")
-        self.env_info.PYTHONPATH.append(os.path.join(self.package_folder, "python"))
         self.env_info.CMAKE_PREFIX_PATH.append(
             os.path.join(self.package_folder, "lib", "cmake")
         )