+++ python/aswfdocker/tests/test_builder.py
@@ -22,7 +22,7 @@ class TestBuilder(unittest.TestCase):
             repo_uri="notauri", source_branch="testing", aswf_version="2019.123"
         )
 
-    def test_package_blosc_2019_dict(self):
+    def test_package_imath_2019_dict(self):
         b = builder.Builder(
             self.build_info,
             groupinfo.GroupInfo(
@@ -32,25 +32,25 @@ class TestBuilder(unittest.TestCase):
                 targets=[],
             ),
         )
-        blosc_version = list(
-            index.Index().iter_versions(constants.ImageType.PACKAGE, "blosc")
+        imath_version = list(
+            index.Index().iter_versions(constants.ImageType.PACKAGE, "imath")
         )[0]
         baked = b.make_bake_dict(False, False, False, False)
         self.assertEqual(
-            baked["target"]["ci-package-blosc-2019"]["tags"],
+            baked["target"]["ci-package-imath-2019"]["tags"],
             [
-                f"{constants.DOCKER_REGISTRY}/aswflocaltesting/ci-package-blosc:2019",
-                f"{constants.DOCKER_REGISTRY}/aswflocaltesting/ci-package-blosc:{blosc_version}",
-                f"{constants.DOCKER_REGISTRY}/aswflocaltesting/ci-package-blosc:latest",
-                f"{constants.DOCKER_REGISTRY}/aswflocaltesting/ci-package-blosc:2019-1.5.0",
+                f"{constants.DOCKER_REGISTRY}/aswflocaltesting/ci-package-imath:2019",
+                f"{constants.DOCKER_REGISTRY}/aswflocaltesting/ci-package-imath:{imath_version}",
+                f"{constants.DOCKER_REGISTRY}/aswflocaltesting/ci-package-imath:latest",
+                f"{constants.DOCKER_REGISTRY}/aswflocaltesting/ci-package-imath:2019-2.3.0",
             ],
         )
         self.assertEqual(
-            baked["target"]["ci-package-blosc-2019"]["args"]["ASWF_VERSION"],
-            blosc_version,
+            baked["target"]["ci-package-imath-2019"]["args"]["ASWF_VERSION"],
+            imath_version,
         )
         self.assertEqual(
-            baked["target"]["ci-package-blosc-2019"]["dockerfile"],
+            baked["target"]["ci-package-imath-2019"]["dockerfile"],
             "packages/vfx1/Dockerfile",
         )